Fixed most picoc warnings

+++ Dependencies/picoc/clibrary.cpp
@@ -2,11 +2,11 @@
 #include "interpreter.h"
 #include <iostream>
 
+
  /* endian-ness checking */
 static const int __ENDIAN_CHECK__ = 1;
 static int BigEndian;
 static int LittleEndian;
-
 
 /* global initialisation for libraries */
 void LibraryInit(Picoc* pc)
@@ -24,6 +24,7 @@ void LibraryInit(Picoc* pc)
     VariableDefinePlatformVar(pc, NULL, (char*)"LITTLE_ENDIAN", &pc->IntType, (union AnyValue*)&LittleEndian, FALSE);
 }
 
+
 /* add a library */
 void LibraryAdd(Picoc* pc, struct Table* GlobalTable, const char* LibraryName, struct LibraryFunction* FuncList)
 {
@@ -38,7 +39,7 @@ void LibraryAdd(Picoc* pc, struct Table* GlobalTable, const char* LibraryName, s
     /* read all the library definitions */
     for (Count = 0; FuncList[Count].Prototype != NULL; Count++)
     {
-        Tokens = LexAnalyse(pc, IntrinsicName, FuncList[Count].Prototype, strlen((char*)FuncList[Count].Prototype), NULL);
+        Tokens = LexAnalyse(pc, IntrinsicName, FuncList[Count].Prototype, (int)strlen((char*)FuncList[Count].Prototype), NULL);
         LexInitParser(&Parser, pc, FuncList[Count].Prototype, Tokens, IntrinsicName, TRUE, FALSE);
         TypeParse(&Parser, &ReturnType, &Identifier, NULL);
         NewValue = ParseFunctionDefinition(&Parser, ReturnType, Identifier);
